requirements.txt
requests
notify-py
pyxdg
